[REFACTOR]: split contextProvider in folder page
