+++ src/components/folder/Modal/Modal.style.ts
@@ -8,7 +8,7 @@ export const ModalBackground = styled.div`
   width: 100%;
   height: 100%;
   background-color: rgba(0, 0, 0, 0.4);
-  z-index: 10;
+  z-index: 1000;
 `;
 
 export const ModalWrapper = styled.div`
